Camera and cleanup Reimplemented the camera system. Did some general cleanup:     Cleared some console logs.     Created a components and a systems folders and moved files around.
