+++ src/enums/COMPONENTS.js
@@ -1,6 +1,6 @@
-import Component_Place from "../engine/ecs/Component_Place";
-import Component_Movement from "../engine/ecs/Component_Movement";
-import Component_Bounding from "../engine/ecs/Component_Bounding";
+import Component_Place from "../engine/ecs/components/Component_Place";
+import Component_Movement from "../engine/ecs/components/Component_Movement";
+import Component_Bounding from "../engine/ecs/components/Component_Bounding";
 
 const COMPONENTS = {
   PLACE: Component_Place,
