+++ src/redux/slices/inputsSlice.js
@@ -12,7 +12,6 @@ export const inputsState = createSlice({
   initialState: initialControls,
   reducers: {
     setInputToTrue: (state, action) => {
-      console.log(action.payload);
       state[action.payload] = true;
     },
     setInputToFalse: (state, action) => {
